requirements.txt
discord.py
python-dotenv
solana
solders
pybase58
websockets
pandas
tabulate
